Fix border painting, add border example

+++ core/src/elements/border.hpp
@@ -10,7 +10,7 @@
 namespace aardvark::elements {
 
 struct BorderSide {
-    int width = 0;
+    float width = 0;
     SkColor color;
 };
